move related component files & moved tests to __tests__ folder
