--- src/components/DependencyInfoFromNpm/index.js
@@ -1,3 +1,0 @@
-// @flow
-export { default } from './DependencyInfoFromNpm';
-export type { NpmResult } from './DependencyInfoFromNpm';
