+++ src/components/DependencyManagementPane/DeleteDependencyButton.js
