+++ src/components/DependencyManagementPane/DependencyInfoFromNpm.js
