+++ src/components/DependencyManagementPane/DependencyUpdateRow.js
