+++ src/components/DependencyManagementPane/__tests__/AddDependencySearchBox.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import AddDependencySearchBox from './AddDependencySearchBox';
+import AddDependencySearchBox from '../AddDependencySearchBox';
 
 describe('AddDependencySearchBox component', () => {
   const wrapper = shallow(<AddDependencySearchBox />);
