--- src/components/DependencyUpdateRow/index.js
@@ -1,2 +1,0 @@
-// @flow
-export { default } from './DependencyUpdateRow';
